refactor: Consolidate several 'mana'-related utilities
