--- packages/@haiku/core/src/helpers/cssMatchById.ts
@@ -1,14 +1,0 @@
-/**
- * Copyright (c) Haiku 2016-2018. All rights reserved.
- */
-
-import objectPath from './objectPath';
-
-export default function matchById (node, id, options) {
-  const attributes = objectPath(node, options.attributes);
-  if (attributes) {
-    if (attributes.id === id) {
-      return true;
-    }
-  }
-}
