+++ packages/@haiku/core/src/helpers/parseCssTransformString.ts
@@ -5,10 +5,13 @@
 import Layout3D from './../Layout3D';
 import cssMat4 from './../vendor/css-mat4';
 import composedTransformsToTimelineProperties from './composedTransformsToTimelineProperties';
-import MathUtils from './MathUtils';
 import parseCssValueString from './parseCssValueString';
 
-function separate (str) {
+function degreesToRadians (d: number): number {
+  return d * Math.PI / 180;
+}
+
+function separate (str: string) {
   const bits = str.split('(');
   const type = bits[0];
   return {
@@ -55,15 +58,15 @@ export default function parseCssTransformString (inStr) {
       // 1D
       case 'rotatex':
         layout.rotate[0] =
-          spec.values[0].unit === 'deg' ? MathUtils.degreesToRadians(spec.values[0].value) : spec.values[0].value;
+          spec.values[0].unit === 'deg' ? degreesToRadians(spec.values[0].value) : spec.values[0].value;
         break;
       case 'rotatey':
         layout.rotate[1] =
-          spec.values[0].unit === 'deg' ? MathUtils.degreesToRadians(spec.values[0].value) : spec.values[0].value;
+          spec.values[0].unit === 'deg' ? degreesToRadians(spec.values[0].value) : spec.values[0].value;
         break;
       case 'rotatez':
         layout.rotate[2] =
-          spec.values[0].unit === 'deg' ? MathUtils.degreesToRadians(spec.values[0].value) : spec.values[0].value;
+          spec.values[0].unit === 'deg' ? degreesToRadians(spec.values[0].value) : spec.values[0].value;
         break;
       case 'translatex':
         layout.translate[0] = spec.values[0].value;
@@ -85,17 +88,17 @@ export default function parseCssTransformString (inStr) {
         break;
       case 'skewx':
         layout.skew[0] =
-          spec.values[0].unit === 'deg' ? MathUtils.degreesToRadians(spec.values[0].value) : spec.values[0].value;
+          spec.values[0].unit === 'deg' ? degreesToRadians(spec.values[0].value) : spec.values[0].value;
         break;
       case 'skewy':
         layout.skew[1] =
-          spec.values[0].unit === 'deg' ? MathUtils.degreesToRadians(spec.values[0].value) : spec.values[0].value;
+          spec.values[0].unit === 'deg' ? degreesToRadians(spec.values[0].value) : spec.values[0].value;
         break;
 
       // 2D
       case 'rotate':
         layout.rotate[2] =
-          spec.values[0].unit === 'deg' ? MathUtils.degreesToRadians(spec.values[0].value) : spec.values[0].value;
+          spec.values[0].unit === 'deg' ? degreesToRadians(spec.values[0].value) : spec.values[0].value;
         break;
       case 'scale':
         layout.scale[0] = spec.values[0].value;
@@ -103,10 +106,10 @@ export default function parseCssTransformString (inStr) {
         break;
       case 'skew':
         layout.skew[0] =
-          spec.values[0].unit === 'deg' ? MathUtils.degreesToRadians(spec.values[0].value) : spec.values[0].value;
+          spec.values[0].unit === 'deg' ? degreesToRadians(spec.values[0].value) : spec.values[0].value;
         if (spec.values[1]) {
           layout.skew[1] =
-            spec.values[1].unit === 'deg' ? MathUtils.degreesToRadians(spec.values[1].value) : spec.values[1].value;
+            spec.values[1].unit === 'deg' ? degreesToRadians(spec.values[1].value) : spec.values[1].value;
         }
         break;
       case 'translate':
@@ -132,7 +135,7 @@ export default function parseCssTransformString (inStr) {
           }
 
           if (axisSpec.unit === 'deg') {
-            return MathUtils.degreesToRadians(axisSpec.value);
+            return degreesToRadians(axisSpec.value);
           }
 
           return axisSpec.value;
