+++ packages/haiku-serialization/src/bll/Bytecode.js
@@ -3,7 +3,7 @@ const clone = require('lodash.clone')
 const cloneDeepWith = require('lodash.clonedeepwith')
 const merge = require('lodash.merge')
 const BaseModel = require('./BaseModel')
-const xmlToMana = require('@haiku/core/lib/helpers/xmlToMana').default
+const {xmlToMana} = require('@haiku/core/lib/HaikuNode')
 const convertManaLayout = require('@haiku/core/lib/layout/convertManaLayout').default
 const expressionToRO = require('@haiku/core/lib/reflection/expressionToRO').default
 const reifyRO = require('@haiku/core/lib/reflection/reifyRO').default
