+++ packages/haiku-serialization/src/bll/File.js
@@ -1,7 +1,7 @@
 const fse = require('fs-extra')
 const {debounce} = require('lodash')
 const path = require('path')
-const xmlToMana = require('@haiku/core/lib/helpers/xmlToMana').default
+const {xmlToMana} = require('@haiku/core/lib/HaikuNode')
 const objectToRO = require('@haiku/core/lib/reflection/objectToRO').default
 const BaseModel = require('./BaseModel')
 const logger = require('./../utils/LoggerInstance')
